Fix resolution errors in declarations with type and value
Prior to this commit, the resolver directly compared the type
of a declaration's type expression to the type of its
initializer. This doesn't work, since the initial expression
is a value, and not a type, while the type expression is.
Thus, `getTypeForDecl` should appropriately adjust the declared
type's kind to match the declaration's kind before comparing
it against the type of the initializer expression.