Allow Context to override chplenv
This commit also makes getChplEnv, the standalone function, generic over
the input map type that specifies initial environment variables. This
is helpful because the production compiler likes to use `map` and
`const char*`, whereas it makes more sense for us to provide an API
based on `unordered_map` and `std::string`. The templating makes both
possible with the same implementation.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>